rename: exception, output > config 이동 외 파일명 수정
