+++ src/main/java/lotto/Application.java
@@ -1,7 +1,9 @@
 package lotto;
 
+import lotto.config.AppConfig;
+
 public class Application {
     public static void main(String[] args) {
-        ApplicationContext.init();
+        AppConfig.init();
     }
 }
