+++ src/main/java/lotto/config/exception/InputException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.config.exception;
 
 public class InputException extends IllegalArgumentException{
     public InputException(ExceptionType exceptionType){
