+++ src/main/java/lotto/domain/Wallet.java
@@ -1,12 +1,12 @@
 package lotto.domain;
 
-import static lotto.ApplicationContext.getDataModel;
+import static lotto.config.AppConfig.getDataModel;
 
+import lotto.config.exception.ExceptionType;
+import lotto.config.exception.InputException;
+import lotto.config.output.MessageType;
+import lotto.config.output.OutputMessage;
 import lotto.domain.constant.LottoConstant;
-import lotto.exception.ExceptionType;
-import lotto.exception.InputException;
-import lotto.output.MessageType;
-import lotto.output.OutputMessage;
 
 public class Wallet {
     private final int wallet;
